refactor: relocate misplaced helpers to their natural layers
D1: Split engines/_helpers.py into _observed.py (effective-params +
request-metrics extraction), _cuda.py (peak-memory/warmup/cleanup), and
_errors.py (OOM/import helpers); move compute_cv to domain/metrics.py.
harness/warmup.py now imports compute_cv from domain (harness->domain).

D2: Move detect_cuda_version_with_source from domain/environment.py to
device/environment.py (Pydantic models stay in domain). Drop the now-unused
subprocess/importlib.util/logging imports from domain/environment.py.

D3: Extract the preflight display (build_preflight_panel,
format_preflight_summary, _resolve_energy_display, _ENERGY_SAMPLER_NAMES)
out of config/grid.py into cli/_preflight_display.py. count_sweep_structure
stays in grid.py. Panel/summary tests moved to tests/unit/cli/test_preflight_display.py.

D4: cli/run.py uses api.study_dir_name (re-exported through the api facade
to respect the cli-boundary contract) instead of re-deriving the study dir
name inline. run_study/_run gain an optional preresolved kwarg so the CLI's
preflight result is reused instead of re-resolving runners inside _run.
