+++ src/llenergymeasure/domain/environment.py
@@ -4,14 +4,9 @@ Captures the hardware and software environment at experiment time,
 enabling post-hoc analysis of environmental factors affecting measurements.
 """
 
-import importlib.util
-import logging
-import subprocess
 from datetime import datetime
 
 from pydantic import BaseModel, Field
-
-logger = logging.getLogger(__name__)
 
 
 class GPUEnvironment(BaseModel):
@@ -155,60 +150,3 @@ class EnvironmentSnapshot(BaseModel):
     tool_version: str
     cuda_version: str | None = None
     cuda_version_source: str | None = None  # "torch" | "version_txt" | "nvcc" | None
-
-
-# ---------------------------------------------------------------------------
-# CUDA version detection - multi-source fallback chain
-# ---------------------------------------------------------------------------
-
-
-def detect_cuda_version_with_source() -> tuple[str | None, str | None]:
-    """Detect the CUDA version using a fallback chain.
-
-    Returns:
-        Tuple of (version_string, source_name) where source_name is one of:
-        "torch", "version_txt", "nvcc", or None if detection failed.
-    """
-    # Source 1: torch.version.cuda
-    if importlib.util.find_spec("torch") is not None:
-        try:
-            import torch
-
-            cuda_ver = torch.version.cuda
-            if cuda_ver:
-                return cuda_ver, "torch"
-        except Exception:
-            logger.debug("CUDA version: torch source failed", exc_info=True)
-
-    # Source 2: /usr/local/cuda/version.txt or version.json
-    import re
-
-    for version_file in (
-        "/usr/local/cuda/version.txt",
-        "/usr/local/cuda/version.json",
-    ):
-        try:
-            with open(version_file) as f:
-                content = f.read()
-            match = re.search(r"(\d+\.\d+)", content)
-            if match:
-                return match.group(1), "version_txt"
-        except Exception:
-            pass
-
-    # Source 3: nvcc --version
-    try:
-        result = subprocess.run(
-            ["nvcc", "--version"],
-            capture_output=True,
-            text=True,
-            timeout=5,
-        )
-        match = re.search(r"release (\d+\.\d+)", result.stdout)
-        if match:
-            return match.group(1), "nvcc"
-    except Exception:
-        logger.debug("CUDA version: nvcc source failed", exc_info=True)
-
-    # Source 4: Give up
-    return None, None
