+++ src/llenergymeasure/engines/tensorrt/plugin.py
@@ -172,7 +172,7 @@ class TensorRTEngine:
         Raises:
             EngineError: If TRT-LLM is not installed or model loading fails.
         """
-        from llenergymeasure.engines._helpers import require_import
+        from llenergymeasure.engines._errors import require_import
 
         _trt_mod = require_import("tensorrt_llm")
         LLM = _trt_mod.LLM
@@ -247,7 +247,7 @@ class TensorRTEngine:
         """
         from tensorrt_llm import SamplingParams
 
-        from llenergymeasure.engines._helpers import warmup_single_token
+        from llenergymeasure.engines._cuda import warmup_single_token
 
         llm, _sampling_params = model
         warmup_single_token(llm, [prompt], SamplingParams, max_tokens=1)
@@ -280,7 +280,7 @@ class TensorRTEngine:
         llm, sampling_params = model
 
         # Reset peak stats before the measurement loop
-        from llenergymeasure.engines._helpers import reset_cuda_peak_memory
+        from llenergymeasure.engines._cuda import reset_cuda_peak_memory
 
         reset_cuda_peak_memory()
 
@@ -295,7 +295,7 @@ class TensorRTEngine:
             outputs = llm.generate(prompts, sampling_params)
             elapsed = time.perf_counter() - t0
         except Exception as e:
-            from llenergymeasure.engines._helpers import raise_engine_error
+            from llenergymeasure.engines._errors import raise_engine_error
 
             raise_engine_error(
                 e,
@@ -304,7 +304,7 @@ class TensorRTEngine:
             )
 
         # Capture peak memory
-        from llenergymeasure.engines._helpers import get_cuda_peak_memory_mb
+        from llenergymeasure.engines._cuda import get_cuda_peak_memory_mb
 
         peak_mb = get_cuda_peak_memory_mb()
 
@@ -333,7 +333,7 @@ class TensorRTEngine:
 
         # Defensive per-request metric capture - RequestOutput.metrics is usually
         # absent in TRT-LLM 0.21.0, so these lists typically come back empty.
-        from llenergymeasure.engines._helpers import extract_request_metrics
+        from llenergymeasure.engines._observed import extract_request_metrics
 
         per_request_latencies_ms, ttft_ms = extract_request_metrics(outputs)
 
@@ -370,7 +370,7 @@ class TensorRTEngine:
         ``_``-prefix allowlist behaviour in
         :func:`extract_observed_params`.
         """
-        from llenergymeasure.engines._helpers import (
+        from llenergymeasure.engines._observed import (
             assemble_observed_params,
             extract_observed_params,
         )
@@ -422,7 +422,7 @@ class TensorRTEngine:
         Args:
             model: Tuple of (llm, sampling_params) from load_model().
         """
-        from llenergymeasure.engines._helpers import cleanup_model
+        from llenergymeasure.engines._cuda import cleanup_model
 
         llm, _sampling_params = model
         cleanup_model(llm)
